Insert und Select jetzt getrennt
