+++ libs/sql-tools/src/index.ts
@@ -1,1 +1,2 @@
-export * from './lib/SqlTools';
+export * from './lib/Select';
+export * from './lib/Insert';
